[FIX] project,{hr,sale}_timesheet: fix update description glitch
Prior to this commit, the spaces in the template were transformed into
non-breaking spaces by the html editor while editing the html field
content.
The unordered list with icons as list item also failed to be compatible
with the editor.

With this commit, we remove the indents in the templates to prevent this
behavior.
It now also uses the css class defined by the editor to generate the check
lists.

closes odoo/odoo#72571

X-original-commit: 8119d6c99f16b18f03df007c6069cc3162656a14
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>