[FIX] sale: improve compute expected date performance
Without this patch, each time the compute reaches the line with `.filtered()`, it needs a new fetch from the database. With too many sale orders, this becomes a bottleneck.

Now, before getting to that line, Odoo knows that it'll need all sale.order.line records, so it can fetch them all at once.

In a production database, loading the sale.order list goes down from 7s to 500ms with this patch.

@Tecnativa TT30390

closes odoo/odoo#72141

X-original-commit: 81d8f5480504addc44a0d9eea0122a3f9b25a8a3
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>