[FIX] website_sale: Make coupons in the cart deletable again
The onClickDeleteProduct handler in the cart widget triggers a value
change in the input field containing the quantity of the sale order
line. This only works if there is such an input field, however these
were removed for non sellable lines (including reward lines) in commit
589352eaa84e92bfde514aced2f3a6e9e9a378fc. Simply restoring the input
field (even if it stays hidden) resolves the problem.

closes odoo/odoo#74222

X-original-commit: 646519c6a7d8e34f3783d51497f9fba6d83555ef
Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>