[FIX] stock: remove boolean widget on signature binary field
Before this commit: when writing something in signture field in stock.picking
and enable signature field in listview throws traceback.

After this commit: when signature field is set and enabling it in list view
will not throw traceback, traceback was generated because boolean field
was applied on Image field, to fix it simply remove boolean widget on
signature field.

task-2570929

closes odoo/odoo#72479

X-original-commit: 2679b6721f74ad26f790285855102a44e30fac71
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>