[FIX] website_event: properly show social share icons on event sidebar
Currently, the social share icons on right sidebar of the event page
are not alligned properly. This is because the rules written from
there are having lower precedence than the generic rules for the
social share snippet.

This commit fixes the issue by providing higher precedence to the
rules specifically written for event's web page.

TaskID-2557077

closes odoo/odoo#72856

X-original-commit: b3d52004172be506c49e4355a21e6c0ec148bc6e
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>